refactor: complete Stage 4 - Booking operations extraction
- Extract BookingOperations class with 3 methods (367 lines)
- Transfer add_booking_to_project, remove_booking_from_project, create_project_with_bookings
- Update project_facade with delegation methods
- Reduce main facade from 796 to 549 lines
- Progress: 10/12 methods completed (83%)
- Maintains 100% backward compatibility through facade pattern
