+++ backend/services/project/operations/__init__.py
@@ -3,7 +3,8 @@
 Contains all project-related operations classes.
 """
 
+from .booking_operations import BookingOperations
 from .crud_operations import CrudOperations
 from .query_operations import QueryOperations
 
-__all__ = ['CrudOperations', 'QueryOperations']
+__all__ = ['BookingOperations', 'CrudOperations', 'QueryOperations']
